requirements.txt
feedparser
google-genai
anthropic
Pillow
moviepy
numpy
edge-tts
